requirements.txt
# Core Streamlit Dependencies
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
typing-extensions>=4.0.0

# Additional Production Dependencies
python-dotenv>=1.0.0
gunicorn>=21.0.0
psutil>=5.9.0

# Development and Testing Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Security and Monitoring
cryptography>=41.0.0
python-dateutil>=2.8.0
